requirements.txt
# Core
fastapi==0.112.2
uvicorn[standard]==0.30.6
sse-starlette==2.1.0

# Pandas + PandasAI
pandas==2.2.2
pandasai>=3.0.0b2
pandasai-litellm==0.0.1
litellm>=1.43.2

# Google Cloud
google-cloud-storage==2.18.2

# Settings
pydantic-settings==2.4.0
python-multipart==0.0.9
